hype_embedder/requirements.txt
langchain
langchain-community
langchain_text_splitters
langchain-ollama
ollama
qdrant-client
python-dotenv
tqdm
marker
marker-pdf
legacy-cgi